Update
Added plater

+++ src/Game/Objects/Entity.java
@@ -1,6 +1,6 @@
 package Game.Objects;
 
 abstract public class Entity {
-    protected int x,y;
+    protected double x,y;
     protected double dx,dy;
 }
